+++ src/app/hook/impl/d3d11_impl.h
@@ -7,7 +7,7 @@ namespace impl
 {
 	namespace d3d11
 	{
-		void init();
+		bool init();
 	}
 }